Reorganise the modules
