+++ fink_client/consumer.py
@@ -21,7 +21,10 @@ import fastavro
 import requests
 from confluent_kafka import KafkaError
 from requests.exceptions import RequestException
-from typing import Any
+
+from fink_client.avroUtils import write_alert
+from fink_client.avroUtils import _get_alert_schema
+from fink_client.avroUtils import _decode_avro_alert
 
 class AlertError(Exception):
     pass
@@ -63,11 +66,11 @@ class AlertConsumer:
         self._consumer.close()
 
     def poll(self, timeout: float = -1) -> (str, dict):
-        """Consume messages from Fink server
+        """ Consume one message from Fink server
 
         Parameters
         ----------
-        timeout: float
+        timeout: float, optional
             maximum time to block waiting for a message
             if not set default is None i.e. wait indefinitely
 
@@ -98,13 +101,12 @@ class AlertConsumer:
         return topic, alert
 
     def consume(self, num_alerts: int = 1, timeout: float = -1) -> list:
-        """Consume and return list of messages
+        """ Consume and return list of messages
 
         Parameters
         ----------
         num_messages: int
             maximum number of messages to return
-
         timeout: float
             maximum time to block waiting for messages
             if not set default is None i.e. wait indefinitely
@@ -126,6 +128,36 @@ class AlertConsumer:
             alerts.append((topic, alert))
 
         return alerts
+
+    def poll_and_write(
+            self, outdir: str, timeout: float = -1,
+            overwrite: bool = False) -> (str, dict):
+        """ Consume one message from Fink server, save alert on disk and
+        return (topic, alert)
+
+        Parameters
+        ----------
+        outdir: str
+            Folder to store the alert. It must exists.
+        timeout: float, optional
+            maximum time to block waiting for messages
+            if not set default is None i.e. wait indefinitely
+        overwrite: bool, optional
+            If True, allow an existing alert to be overwritten.
+            Default is False.
+
+        Returns
+        ----------
+        (topic, alert): tuple(str, dict)
+            returns (None, None) on timeout
+
+        """
+        topic, alert = self.poll(timeout)
+        if topic is not None:
+            # print('Alert written at {}'.format(outdir))
+            write_alert(alert, self._parsed_schema, outdir, overwrite=overwrite)
+
+        return topic, alert
 
     def close(self):
         """Close connection to Fink broker"""
@@ -174,70 +206,3 @@ def _get_kafka_config(config: dict) -> dict:
         kafka_config["bootstrap.servers"] = "{}".format(",".join(fink_servers))
 
     return kafka_config
-
-def _get_alert_schema(schema_path: str = None):
-    """Returns schema for decoding avro alert
-
-    This method downloads the latest schema available on the fink servers
-    or falls back to using a default schema located in dir 'schemas'/
-
-    Parameters
-    ----------
-    schema_path: str, optional
-        a local path where to look for schema,
-        Note that schema doesn't get downloaded from fink servers if schema_path
-        is given
-
-    Returns
-    ----------
-    parsed_schema: dict
-        Dictionary of json format schema for decoding avro alerts from fink
-    """
-    if schema_path is None:
-        # get schema from fink-broker
-        try:
-            print("Getting schema from fink servers...")
-            schema_url = "https://raw.github.com/astrolabsoftware/fink-broker/master/schemas/distribution_schema_0p2.avsc"
-            filename = schema_url.split("/")[-1]
-            r = requests.get(schema_url, timeout=1)
-            schema_path = os.path.abspath(os.path.join(
-                os.path.dirname(__file__), '../schemas/{}'.format(filename)))
-            with open(schema_path, "w") as f:
-                f.write(r.text)
-        except RequestException:
-            schema_path = os.path.abspath(
-                os.path.join(
-                    os.path.dirname(__file__),
-                    '../schemas/distribution_schema_0p2.avsc'
-                )
-            )
-
-            msg = """
-            Could not obtain schema from fink servers
-            Using default schema available at: {}
-            """.format(schema_path)
-            print(msg)
-
-    with open(schema_path) as f:
-        schema = json.load(f)
-
-    return fastavro.parse_schema(schema)
-
-def _decode_avro_alert(avro_alert: io.IOBase, schema: dict) -> Any:
-    """Decodes a file-like stream of avro data
-
-    Parameters
-    ----------
-    avro_alert: io.IOBase
-        a file-like stream with avro encoded data
-
-    schema: dict
-        Dictionary of json format schema to decode avro data
-
-    Returns
-    ----------
-    record: Any
-        Record obtained after decoding avro data (typically, dict)
-    """
-    avro_alert.seek(0)
-    return fastavro.schemaless_reader(avro_alert, schema)
